Fix 64-bit MSVC warnings

+++ src/win32/utf-conv.c
@@ -72,10 +72,10 @@ void git__utf8_to_16(wchar_t *dest, size_t length, const char *src)
 
 int git__utf8_to_16(wchar_t * dest, size_t dest_size, const char *src)
 {
-	return MultiByteToWideChar(CP_UTF8, 0, src, -1, dest, dest_size);
+	return MultiByteToWideChar(CP_UTF8, 0, src, -1, dest, (int)dest_size);
 }
 
 int git__utf16_to_8(char *dest, size_t dest_size, const wchar_t *src)
 {
-	return WideCharToMultiByte(CP_UTF8, 0, src, -1, dest, dest_size, NULL, NULL);
+	return WideCharToMultiByte(CP_UTF8, 0, src, -1, dest, (int)dest_size, NULL, NULL);
 }
